pyproject.toml
[project]
name = "meta-threads-multiagent"
version = "0.1.0"
description = "LangGraph-based AI agent for automated Threads posting with web search"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.12"
authors = [{ name = "MetaThreads" }]
keywords = ["threads", "langgraph", "agent", "mcp", "openrouter", "multiagent"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    "langgraph>=0.2",
    "langchain>=0.3",
    "langchain-core>=0.3",
    "openai>=1.0",
    "mcp>=1.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "httpx>=0.27",
    "sse-starlette>=2.0",
    "feedparser>=6.0",
    "langfuse>=2.0",
    "fastmcp>=2.14.2",
    "ddgs>=9.10.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-httpx>=0.35",
    "pytest-cov>=6.0",
    "ruff>=0.8",
    "mypy>=1.13",
]

[project.scripts]
threads-hype-agent = "threads_hype_agent.api.app:main"

[project.urls]
Homepage = "https://github.com/MetaThreads/meta-threads-multiagent"
Repository = "https://github.com/MetaThreads/meta-threads-multiagent"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/threads_hype_agent"]

[tool.ruff]
target-version = "py312"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults
]

[tool.ruff.lint.isort]
known-first-party = ["threads_hype_agent"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
plugins = ["pydantic.mypy"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --cov=threads_hype_agent --cov-report=term-missing --cov-report=html"

[tool.coverage.run]
source = ["src/threads_hype_agent"]
branch = true
omit = [
    "*/tests/*",
    "*/__init__.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.:",
    "@abstractmethod",
    "\\.\\.\\.",
]
show_missing = true
precision = 2

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
]
